Aggregations and counter field support.
No longer offer limited aggregation support counter fields.
This extends the counter field support to all aggregation.

Closes #93539
