+++ modules/aggregations/src/yamlRestTest/resources/rest-api-spec/test/aggregations/sum_metric.yml
@@ -175,3 +175,42 @@ setup:
             the_string_sum:
               sum:
                 field: string_field
+
+---
+"Counter field":
+  - skip:
+      version: " - 8.6.99"
+      reason: "counter field support added in 8.7"
+
+  - do:
+      indices.create:
+        index: myindex
+        body:
+          mappings:
+            properties:
+              counter_field:
+                type : long
+                time_series_metric: counter
+
+  - do:
+      bulk:
+        refresh: true
+        body:
+          - create:
+              _index: myindex
+          - counter_field: 50
+          - create:
+              _index: myindex
+          - counter_field: 100
+
+  - do:
+      search:
+        rest_total_hits_as_int: true
+        index: myindex
+        body:
+          aggs:
+            the_counter_sum:
+              sum:
+                field: counter_field
+  - match: { hits.total: 2 }
+  - match: { aggregations.the_counter_sum.value: 150.0 }
