+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/histogram/VariableWidthHistogramAggregatorFactory.java
@@ -14,13 +14,11 @@ import org.elasticsearch.search.aggregations.AggregatorFactory;
 import org.elasticsearch.search.aggregations.CardinalityUpperBound;
 import org.elasticsearch.search.aggregations.support.AggregationContext;
 import org.elasticsearch.search.aggregations.support.CoreValuesSourceType;
-import org.elasticsearch.search.aggregations.support.TimeSeriesValuesSourceType;
 import org.elasticsearch.search.aggregations.support.ValuesSourceAggregatorFactory;
 import org.elasticsearch.search.aggregations.support.ValuesSourceConfig;
 import org.elasticsearch.search.aggregations.support.ValuesSourceRegistry;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
 
 public class VariableWidthHistogramAggregatorFactory extends ValuesSourceAggregatorFactory {
@@ -28,7 +26,7 @@ public class VariableWidthHistogramAggregatorFactory extends ValuesSourceAggrega
     public static void registerAggregators(ValuesSourceRegistry.Builder builder) {
         builder.register(
             VariableWidthHistogramAggregationBuilder.REGISTRY_KEY,
-            List.of(CoreValuesSourceType.NUMERIC, TimeSeriesValuesSourceType.COUNTER),
+            CoreValuesSourceType.ALL_NUMERIC,
             VariableWidthHistogramAggregator::new,
             true
         );
