+++ server/src/main/java/org/elasticsearch/search/aggregations/metrics/SumAggregatorFactory.java
@@ -19,7 +19,6 @@ import org.elasticsearch.search.aggregations.support.ValuesSourceConfig;
 import org.elasticsearch.search.aggregations.support.ValuesSourceRegistry;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
 
 class SumAggregatorFactory extends ValuesSourceAggregatorFactory {
@@ -41,12 +40,7 @@ class SumAggregatorFactory extends ValuesSourceAggregatorFactory {
     }
 
     static void registerAggregators(ValuesSourceRegistry.Builder builder) {
-        builder.register(
-            SumAggregationBuilder.REGISTRY_KEY,
-            List.of(CoreValuesSourceType.NUMERIC, CoreValuesSourceType.DATE, CoreValuesSourceType.BOOLEAN),
-            SumAggregator::new,
-            true
-        );
+        builder.register(SumAggregationBuilder.REGISTRY_KEY, CoreValuesSourceType.ALL_NUMERIC_LENIENT, SumAggregator::new, true);
     }
 
     @Override
