+++ x-pack/plugin/analytics/src/main/java/org/elasticsearch/xpack/analytics/rate/RateAggregatorFactory.java
@@ -16,7 +16,6 @@ import org.elasticsearch.search.aggregations.CardinalityUpperBound;
 import org.elasticsearch.search.aggregations.LeafBucketCollector;
 import org.elasticsearch.search.aggregations.support.AggregationContext;
 import org.elasticsearch.search.aggregations.support.CoreValuesSourceType;
-import org.elasticsearch.search.aggregations.support.TimeSeriesValuesSourceType;
 import org.elasticsearch.search.aggregations.support.ValuesSourceAggregatorFactory;
 import org.elasticsearch.search.aggregations.support.ValuesSourceConfig;
 import org.elasticsearch.search.aggregations.support.ValuesSourceRegistry;
@@ -67,7 +66,7 @@ class RateAggregatorFactory extends ValuesSourceAggregatorFactory {
         );
         builder.register(
             RateAggregationBuilder.REGISTRY_KEY,
-            Collections.singletonList(TimeSeriesValuesSourceType.COUNTER),
+            Collections.singletonList(CoreValuesSourceType.COUNTER),
             TimeSeriesRateAggregator::new,
             true
         );
